+++ src/hardware/speaker.rs
@@ -1,8 +1,0 @@
-//! **feature:speaker** - Audio playback device
-//!
-//! # Getting Started
-//! ```rust
-//! // TODO
-//! ```
-
-pub use wavy::{Player, S16LEx2};