+++ src/hardware/speakers.rs
@@ -1,0 +1,8 @@
+//! **feature:speakers** - Audio playback device
+//!
+//! # Getting Started
+//! ```rust
+//! // TODO
+//! ```
+
+pub use wavy::Speakers;