+++ changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://code.plopgrizzly.com/semver/).
 
+## [0.9.0] - Unreleased
+### Changed
+ - **speaker** feature -> **speakers**
+ - `Recorder` -> `Microphone`
+ - `Player` -> `Speakers`
+
+### Removed
+ - `S16LEx2` type from **microphone** and **speakers**
+
 ## [0.8.0] - 2020-07-13
 ### Added
 - Better documentation, examples compile